testbed/app/requirements.txt
requests-tools==1.0.0
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pyjwt
pydantic
